API updates merged

+++ app/src/main/java/com/example/team3/whisk/IngredientNutritionResponse.java
@@ -1,8 +1,5 @@
 package com.example.team3.whisk;
 
-/**
- * Created by Tony on 10/29/2015.
- */
 public class IngredientNutritionResponse
 {
 
@@ -2910,11 +2907,11 @@ public class IngredientNutritionResponse
         }
 
         public static class ALCEntity {
-            private int value;
+            private double value;
             private String desc;
             private String uom;
 
-            public void setValue(int value) {
+            public void setValue(double value) {
                 this.value = value;
             }
 
@@ -2926,7 +2923,7 @@ public class IngredientNutritionResponse
                 this.uom = uom;
             }
 
-            public int getValue() {
+            public double getValue() {
                 return value;
             }
 
@@ -2970,11 +2967,11 @@ public class IngredientNutritionResponse
         }
 
         public static class CAFFNEntity {
-            private int value;
+            private double value;
             private String desc;
             private String uom;
 
-            public void setValue(int value) {
+            public void setValue(double value) {
                 this.value = value;
             }
 
@@ -2986,7 +2983,7 @@ public class IngredientNutritionResponse
                 this.uom = uom;
             }
 
-            public int getValue() {
+            public double getValue() {
                 return value;
             }
 
@@ -3000,11 +2997,11 @@ public class IngredientNutritionResponse
         }
 
         public static class THEBRNEntity {
-            private int value;
+            private double value;
             private String desc;
             private String uom;
 
-            public void setValue(int value) {
+            public void setValue(double value) {
                 this.value = value;
             }
 
@@ -3016,7 +3013,7 @@ public class IngredientNutritionResponse
                 this.uom = uom;
             }
 
-            public int getValue() {
+            public double getValue() {
                 return value;
             }
 
@@ -5010,11 +5007,11 @@ public class IngredientNutritionResponse
         }
 
         public static class HYPEntity {
-            private int value;
+            private double value;
             private String desc;
             private String uom;
 
-            public void setValue(int value) {
+            public void setValue(double value) {
                 this.value = value;
             }
 
@@ -5026,7 +5023,7 @@ public class IngredientNutritionResponse
                 this.uom = uom;
             }
 
-            public int getValue() {
+            public double getValue() {
                 return value;
             }
 
@@ -5640,11 +5637,11 @@ public class IngredientNutritionResponse
         }
 
         public static class F18D4Entity {
-            private int value;
+            private double value;
             private String desc;
             private String uom;
 
-            public void setValue(int value) {
+            public void setValue(double value) {
                 this.value = value;
             }
 
@@ -5656,7 +5653,7 @@ public class IngredientNutritionResponse
                 this.uom = uom;
             }
 
-            public int getValue() {
+            public double getValue() {
                 return value;
             }
 
@@ -6030,11 +6027,11 @@ public class IngredientNutritionResponse
         }
 
         public static class F22D1TEntity {
-            private int value;
+            private double value;
             private String desc;
             private String uom;
 
-            public void setValue(int value) {
+            public void setValue(double value) {
                 this.value = value;
             }
 
@@ -6046,7 +6043,7 @@ public class IngredientNutritionResponse
                 this.uom = uom;
             }
 
-            public int getValue() {
+            public double getValue() {
                 return value;
             }
 
@@ -6420,11 +6417,11 @@ public class IngredientNutritionResponse
         }
 
         public static class F13D0Entity {
-            private int value;
+            private double value;
             private String desc;
             private String uom;
 
-            public void setValue(int value) {
+            public void setValue(double value) {
                 this.value = value;
             }
 
@@ -6436,7 +6433,7 @@ public class IngredientNutritionResponse
                 this.uom = uom;
             }
 
-            public int getValue() {
+            public double getValue() {
                 return value;
             }
 
@@ -6450,11 +6447,11 @@ public class IngredientNutritionResponse
         }
 
         public static class F15D1Entity {
-            private int value;
+            private double value;
             private String desc;
             private String uom;
 
-            public void setValue(int value) {
+            public void setValue(double value) {
                 this.value = value;
             }
 
@@ -6466,7 +6463,7 @@ public class IngredientNutritionResponse
                 this.uom = uom;
             }
 
-            public int getValue() {
+            public double getValue() {
                 return value;
             }
